Implement new thread pool
This is v.3 of the new thread pool implementation. It is somewhat simpler than the current thread
  pool. I ran some performance benchmarks and it seems to have performance equal to or better than
  the current thread pool.

Some notes:

- Tasks are submitted to ThreadPool::async() (for which ThreadPool::execute() is an alias). This
  function has the same interface as std::async() -- meaning it is not necessary to submit a
  packaged task (nor a std::function<void()>).
- When submitting a task, a C++ std::future is returned, rather than a custom type.
- Tasks are put into a thread safe producer-consumer queue, which manages synchronization between
  submitters of work and worker threads.  The producer-consumer queue (also part of this PR) is
  templated on the type of underlying container.  The default is std::deque, which is what is used
  by the thread pool.
- The threadpool has exactly as many threads as are requested by init().
- The default pool size is now set to hardware concurrency rather than to 1 (which in the current
  thread pool is actually zero).
- I added a small random delay to the bodies of the functions in the recursion tests to try to
  create more chaotic behavior. This does increase the time it takes the unit test for thread pool
  to run.
- The new thread pool properly handles exceptions thrown in the tasks being executed. The exceptions
  are attached to the shared state of the future and will have to be caught when future::get() or
  future::wait() are called. Tests have been added to unit_thread_pool for exceptions.
- If the preprocessor macro LEGACY_THREAD_POOL is defined, the legacy thread pool rather than the
  new thread pool will be compiled. This macro can be set in bootstrap with
  --enable-legacy-thread-pool or passed as a cmake option with -DTILEDB_LEGACY_THREAD_POOL=ON. This
  will propagate to building the tiledb library, the tiledb unit test, and the threadpool unit test.
- A testing matrix for legacy thread pool ON or OFF has been added to unit-test-runs.yml so that the
  legacy thread pool will have at least a sanity check test.

---
TYPE: IMPROVEMENT
DESC: A new thread pool with modern C++ compatible API and exception-safe behavior.
