+++ .github/workflows/unit-test-runs.yml
@@ -18,8 +18,9 @@ jobs:
     strategy:
       matrix:
         os: [macos-latest, ubuntu-latest]
+        tp: [ON, OFF]
       fail-fast: false
-    name: Build - ${{ matrix.os }}
+    name: Build - ${{ matrix.os }} (${{ matrix.tp }})
     steps:
       - uses: actions/checkout@v2
       - name: 'Print env'
@@ -52,7 +53,8 @@ jobs:
             -DTILEDB_AZURE=${TILEDB_AZURE}\
             -DTILEDB_GCS=${TILEDB_GCS} \
             -DTILEDB_S3=${TILEDB_S3} \
-            -DTILEDB_SERIALIZATION=${TILEDB_SERIALIZATION}
+            -DTILEDB_SERIALIZATION=${TILEDB_SERIALIZATION} \
+            -DTILEDB_LEGACY_THREAD_POOL=${{ matrix.tp }}
           make -j4
           # Build all unit tests
           make -C tiledb tests -j4
