+++ cmake/TileDB-Superbuild.cmake
@@ -42,6 +42,7 @@ set(INHERITED_CMAKE_ARGS
   -DTILEDB_TOOLS=${TILEDB_TOOLS}
   -DTILEDB_SERIALIZATION=${TILEDB_SERIALIZATION}
   -DTILEDB_ARROW_TESTS=${TILEDB_ARROW_TESTS}
+  -DTILEDB_LEGACY_THREAD_POOL=${TILEDB_LEGACY_THREAD_POOL}
   -DTILEDB_INSTALL_LIBDIR=${TILEDB_INSTALL_LIBDIR}
   -DCMAKE_OSX_ARCHITECTURES=${CMAKE_OSX_ARCHITECTURES}
 )
