+++ test/CMakeLists.txt
@@ -181,7 +181,6 @@ set(TILEDB_UNIT_TEST_SOURCES
   src/unit-SubarrayPartitioner-dense.cc
   src/unit-SubarrayPartitioner-error.cc
   src/unit-SubarrayPartitioner-sparse.cc
-  src/unit-threadpool.cc
   src/unit-Tile.cc
   src/unit-TileDomain.cc
   src/unit-uri.cc
