+++ tiledb/sm/query/reader.cc
@@ -825,7 +825,7 @@ void Reader::compute_fixed_cs_partitions(
     return;
   }
 
-  const int num_copy_threads =
+  const auto num_copy_threads =
       storage_manager_->compute_tp()->concurrency_level();
 
   // Calculate the partition sizes.
@@ -1007,7 +1007,7 @@ void Reader::compute_var_cs_partitions(
     return;
   }
 
-  const int num_copy_threads =
+  const auto num_copy_threads =
       storage_manager_->compute_tp()->concurrency_level();
 
   // Calculate the partition range.
